requirements.txt
pycrypto
lxml
tensorflow
sklearn
numpy